cosmos-python/requirements.txt
azure-cosmos>=4.0.0b6
python-dotenv>=0.13.0
